requirements.txt
django
djangorestframework
psycopg2-binary
dj-database-url